fix routeString tests
